fixed 1

+++ ChatTutor/vectordatabase.py
@@ -103,7 +103,11 @@ class VectorDatabase:
         )
 
     def delete_datasource_chroma(self, collection_name):
-        self.client.delete_collection(name=collection_name)
+        collections = self.client.list_collections()
+        coll_names = [coll.name for coll in collections]
+        print(coll_names, collection_name)
+        if collection_name in coll_names:
+            self.client.delete_collection(name=collection_name)
 
     def add_texts(self, texts: List[Text]):
         """Adding texts to the database based on the database provider
